pyproject.toml
[project]
name = "checkin-box"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
]
